dsa-indexer(graph-move,mechanical_provable): split prefill CUDA graph logic
